+++ spec/array_include_methods_spec.rb
@@ -211,16 +211,50 @@ describe 'ArrayIncludeMethods' do
   end
   
   describe '#array_diff_indexes' do
+    it 'returns [1, 2]' do
+      expect(array1.array_intersection_indexes(array2)).to eq([1, 2])
+    end
+
+    it 'returns [0]' do
+      expect(array1.array_intersection_indices(array15)).to eq([0])
+    end
+
+    it 'returns [1, 3]' do
+      expect(array1.array_intersection_indexes(array2a)).to eq([1, 3])
+    end
+
+    it 'returns [3]' do
+      expect(array1.array_intersection_indexes(array8)).to eq([3])
+    end
+
+    it 'returns []' do
+      expect(array1.array_intersection_indexes(array5)).to eq([])
+    end
+
+    it 'returns []' do
+      expect(array4.array_intersection_indexes(array5)).to eq([])
+    end
+
+    it 'returns [] for nil array' do
+      expect(array1.array_intersection_indexes(array6)).to eq([])
+    end
+  end
+  
+  describe '#array_diff_indexes' do
     it 'returns [0, 3]' do
       expect(array1.array_diff_indexes(array2)).to eq([0, 3])
     end
 
     it 'returns [1, 2, 3]' do
-      expect(array1.array_diff_indexes(array15)).to eq([1, 2, 3])
+      expect(array1.array_diff_indices(array15)).to eq([1, 2, 3])
     end
 
     it 'returns [0, 2]' do
       expect(array1.array_diff_indexes(array2a)).to eq([0, 2])
+    end
+
+    it 'returns [0, 1, 2]' do
+      expect(array1.array_diff_indexes(array8)).to eq([0, 1, 2])
     end
 
     it 'returns [0, 1, 2, 3]' do
@@ -231,8 +265,8 @@ describe 'ArrayIncludeMethods' do
       expect(array4.array_diff_indexes(array5)).to eq([])
     end
 
-    it 'returns [] for nil array' do
-      expect(array1.array_diff_indexes(array6)).to eq([])
+    it 'returns [0, 1, 2, 3] for nil array' do
+      expect(array1.array_diff_indexes(array6)).to eq([0, 1, 2, 3])
     end
   end
 end